pipeline/kafka-mq-bridge/requirements.txt
confluent-kafka==2.3.0
aws-msk-iam-sasl-signer-python==1.0.2
stomp.py==8.1.0
